+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] - ReleaseDate
+### Fixed
+* Panic when trying to fit a model to an empty trace
+
 ## [0.4.2] - 2019-03-19
 ### Added
 * Spinner animation during capture, load and model fitting #24